// comonl-comonlwcl/src/app/modules/comonlapi/model/formRicercaAccreditamentoAnagrafiche.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface FormRicercaAccreditamentoAnagrafiche { 
    cf?: string;
    cognome?: string;
    nome?: string;
    cfAzienda?: string;
    getpIva?: string;
    denom?: string;
    annullate?: boolean;
}