// comonl-comonlwcl/src/app/modules/comonlapi/model/pagedResponseRicercaAccreditamentoConsulente.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AnagraficaDelegato } from './anagraficaDelegato';


export interface PagedResponseRicercaAccreditamentoConsulente { 
    list?: Array<AnagraficaDelegato>;
    currentPage?: number;
    totalPages?: number;
    totalElements?: number;
}